+++ lib/src/tooltip/animated_tooltip_multi_layout.dart
@@ -59,7 +59,7 @@ class _AnimatedTooltipMultiLayout extends MultiChildRenderObjectWidget {
   final double gapBetweenContentAndAction;
   final double toolTipSlideEndDistance;
   final Alignment? scaleAlignment;
-  final double screenEdgePadding;
+  final EdgeInsets screenEdgePadding;
   final EdgeInsets targetPadding;
   final Offset showcaseOffset;
   final double targetTooltipGap;